### 2024-06-25 #6929 @editorialbot
[PRE REVIEW]: ASIMTools: A lightweight workflow and simulation manager for reproducible atomistic simulations
**Submitting author:** @mkphuthi (<a href="http://orcid.org/0000-0002-0982-8635">Mgcini Keith Phuthi</a>)
**Repository:** https://github.com/BattModels/asimtools.git
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.0
