### 2024-06-25 #6929 @danielskatz
👋 @mkphuthi - thanks for your submission.

Can you please add the countries to the author affiliations?

Also, JOSS papers are requested to be around 1000 words at most, and this is 2000, so you will have to reduce the content somewhat. Sometimes, there are parts of the paper that either duplicate documentation in the repo, or that could move into documentation in the repo (in this case, perhaps one or both of the examples).

Finally, you need to add a references header at the end of the .md file, so that when the references are added to the PDF, they have a header in front of them.

After you make changes in your .md file, you can then use the command `@editorialbot generate pdf` to make a new PDF and recount the word length. editorialbot commands need to be the first entry in a new comment.
