### 2024-06-25 eff6ad6
paper
