+++ paper.md
@@ -19,9 +19,9 @@ authors:
     equal-contrib: false
     affiliation: "1" # (Multiple affiliations must be quoted)
 affiliations:
- - name: University of Michigan
+ - name: University of Michigan, United States
    index: 1
- - name: Carnegie Mellon University
+ - name: Carnegie Mellon University, United States
    index: 2
 date: 17 June 2024
 bibliography: paper.bib
@@ -39,25 +39,25 @@ manager for reproducible atomistic simulations on Unix-based systems. Within
 the framework, simulations can be transferred across computing environments,
 DFT codes, interatomic potentials and atomic structures. By using in-built or
 user-defined python modules (called asimmodules) and utilities, users can run
-simulation recipes and automatically scale them on slurm based clusters or
+simulation protocols and automatically scale them on slurm based clusters or
 locally on their console. The core idea is to separate the dependence of the
 atomistic potential/calculator, the computing environment and the simulation
 protocol thereby allowing the same simulation to be run with different
-calculators, structures or on different computers with just a change of one
-parameter in an input file after initial setup. This is increasingly necessary
-as benchmarking Machine Learning Interatomic Potentials has become a core part
-of computational materials science. Input and output files follow a simple
-standard format, usually yaml, providing a simple interface that also acts as a
-record of the parameters used in a simulation without having to edit python
-scripts. The minimal set of requirements means any materials science codes can
-be incorporated into an ASIMTools workflow in a unified way.
+calculators, atomic structures or on different computers with just a change of
+one parameter in an input file after initial setup. This is increasingly
+necessary as benchmarking Machine Learning Interatomic Potentials has become a
+core part of computational materials science. Input and output files follow a
+simple standard format, usually yaml, providing a simple interface that also
+acts as a record of the parameters used in a simulation without having to edit
+python scripts. The minimal set of requirements means any materials science
+codes can be incorporated into an ASIMTools workflow in a unified way.
 
 # Statement of need
 
 Atomic simulations are a key component of modern day materials science in both
 academia and industry. However, simulation protocols and workflows used by
 researchers are typically difficult to transfer to systems using different
-inputs, codes and environments. It often involves rewriting entire scripts in
+inputs, codes and computing environments. It often involves rewriting entire scripts in
 different languages to change from one type of atomistic potential or atomic
 structure to another. This leads to poor reproducibility and inefficient
 transfer of code from one researcher to the next. In addition, there exists a
@@ -75,11 +75,10 @@ community input, these simulation protocols will hopefully converge towards
 best practice and ensure reproducibility of simulation results.
 
 `ASIMTools` is for users interested in performing atomistic calculations on
-UNIX-like operating systems and/or on slurm based High Performance Computing
-clusters. By defining simulation protocols as functions in "asimmodules", they
+UNIX-like operating systems and/or on slurm-based High Performance Computing
+clusters. By defining simulation protocols as "asimmodules", they
 can be easily added to the library of provided asimmodules and iterated on.
-This will allow the community to develop a robust set of shareable simulation
-protocols. The flexibility of ASIMTools allows integration of any kind of
+The flexibility of ASIMTools allows integration of any kind of
[…]
