### 2024-06-26 #6929 @danielskatz
thanks @mkphuthi - we're just having a short discussion about which JOSS track this best fits in, then we'll start the process
