### 2024-07-09 #6929 @HaoZeke
Thanks for the invite @Kevin-Mattheus-Moerman, I'm happy to accept

@mkphuthi thanks for the suggestions, I will start sending out invitations by the end of this week.
