### 2024-07-09 #6929 @HaoZeke
hi @utf @Andrew-S-Rosen @janosh 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

If not, could you recommend any potential reviewers? I was hoping to have your insights because of your work on [atomate](https://github.com/materialsproject/atomate2) and [jobflow](https://github.com/materialsproject/jobflow).
