### 2024-07-09 #6929 @HaoZeke
hi @PavelStishenko @thohamm 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

If not, could you recommend any potential reviewers? I was hoping to have your insights because of your work on ASI woskflows and strucscan..
