### 2024-07-09 #6929 @Andrew-S-Rosen
@HaoZeke: Thank you for reaching out. Unfortunately, I will have to decline as I have too much of a conflict of interest due to https://github.com/Quantum-Accelerators/quacc.
