### 2024-08-06 #6929 @askhl
Hi Rohit,

I'm very sorry, but there is no way I will have time to review.

Best regards
Ask

Am Di., 9. Juli 2024 um 19:54 Uhr schrieb Rohit Goswami <
***@***.***>:

> hi @askhl <https://github.com/askhl> 👋 would you be interested in and
> available to review this JOSS submission? We carry out our checklist-driven
> reviews here in GitHub issues and follow these guidelines:
> joss.readthedocs.io/en/latest/review_criteria.html
>
> If not, could you recommend any potential reviewers? I was hoping to have
> your insights because of your work on ASE and ASE workflows...
>
