### 2024-08-15 001610a
Previously phonopy.save.yaml was shared across different qha calcs, which is not ideal. Also deleted the requirement of phonopy_save_path in qha_properties because it is never used
