### 2024-08-27 3e02086
remove user control over qha_save.yaml path
