### 2024-09-06 BattModels/asimtools#44 @imtambovtcev
[JOSS Review] MAINT: Add a PyPI installation option
The installation instructions are clear; however, adding a PyPI installation option would be a valuable enhancement and improve accessibility for users.

[This suggestion is part of my review in the JOSS submission](https://github.com/openjournals/joss-reviews/issues/7085).
