### 2024-09-06 BattModels/asimtools#45 @imtambovtcev
[JOSS Review] DOC: Emphasize ASIMTools advantages over existing solutions
The paper outlines the capabilities of ASIMTools well, but it would be helpful to further emphasize the specific advantages of this tool over existing solutions. Clarifying what unique contributions ASIMTools offers compared to what already exists would strengthen the documentation. This could involve a more direct explanation of how ASIMTools enhances current workflows and addresses challenges that other tools may not fully resolve.

[This suggestion is part of my review in the JOSS submission](https://github.com/openjournals/joss-reviews/issues/7085).
