### 2024-09-09 #7085 @imtambovtcev
Review of "ASIMTools: A lightweight framework for scalable and reproducible atomic simulations"

Summary:

The manuscript presents ASIMTools, a flexible and lightweight workflow manager designed for reproducible atomic simulations. The tool's ability to handle different atomic structures, computing environments, and simulation protocols with minimal setup is well-explained. The software's potential to simplify and enhance the reproducibility of simulation workflows is evident.

Comments and Suggestions:

Highlighting the Advantages: While the paper effectively outlines the capabilities of ASIMTools, I suggest further emphasizing the specific advantages of this tool over existing solutions. It would be beneficial to clarify in the documentation what unique contributions ASIMTools offers compared to what already exists. This could involve a more direct explanation of how ASIMTools enhances current workflows and addresses challenges that other tools may not fully resolve.

Installation via PyPI: The installation instructions are clear; however, adding a PyPI installation option would be a valuable enhancement.

Conclusion:

I find the manuscript well-prepared, with the suggestions provided to further strengthen the clarity and accessibility of the tool.
