### 2024-09-10 BattModels/asimtools#46 @abhishektiwari
[JOSS Review] Add Github action for running automated tests
Please add Github action for running automated tests. Given tests already exists, it should be very easy to automate the run on every commit or merge.

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/7085)
