### 2024-09-23 BattModels/asimtools#45 @mkphuthi
Dear @imtambovtcev ,

We tried to address this in the paper in the statement of need section:

> There is however no unifying framework that can
> encompass all these tools without significant software development effort.
> Significant effort should not be necessary because while the source of the
> fundamental outputs of atomistic potentials such as energy, forces etc. may
> differ, simulation protocols built on these outputs should converge towards the
> most accurate and computationally efficient. ASIMTools focuses on this last
> aspect by introducing asimmodules which are simply Python functions that act as
> simulation protocols which have no dependence on a specific atomistic potential
> or computational environment or atomic structure.

Is the suggestion that we rephrase more clearly or that we cite other works or that we include this in the docs? Some clarity would be helpful

Thanks.
