### 2024-09-23 2137eb8
Create python-app.yml
Created an action for automated testing
