### 2024-09-23 BattModels/asimtools#46 @mkphuthi
Hi, Thanks for the great suggestion. Tests are now automated on pushes to main.
