### 2024-09-23 BattModels/asimtools#44 @mkphuthi
Hi, Thanks for the suggestion. A PyPi installation option has been added and can be found at [](https://pypi.org/project/asimtools/0.0.0/)
