### 2024-09-24 BattModels/asimtools#45 @imtambovtcev
Dear @mkphuthi,

Thank you for your response. One way to further clarify the advantages of ASIMTools could be to include an example in the documentation that outlines a common problem in atomic simulations and demonstrates how ASIMTools provides a solution. A "problem and solution" format might help illustrate its unique contributions more concretely.

If this approach works for you, it could be a helpful addition.

Best regards,
@imtambovtcev
