+++ .github/workflows/docs.yml
@@ -13,15 +13,15 @@ jobs:
       - uses: actions/setup-python@v5
       - name: Install dependencies
         run: |
-          pip install sphinx sphinx_rtd_theme myst_parser autodoc
+          pip install sphinx sphinx_rtd_theme myst_parser autodoc asimtools
       - name: Sphinx build
         run: |
-          sphinx-build docs _build
+          sphinx-build docs docs_build
       - name: Deploy to GitHub Pages
         uses: peaceiris/actions-gh-pages@v3
         if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
         with:
           publish_branch: gh-pages
           github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: _build/
+          publish_dir: docs_build/
           force_orphan: true
