+++ .github/workflows/docs.yml
@@ -11,6 +11,8 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
+        with:
+          python-version: '3.9'
       - name: Install dependencies
         run: |
           pip install sphinx sphinx_rtd_theme myst_parser autodoc asimtools
