+++ .github/workflows/docs.yml
@@ -20,12 +20,12 @@ jobs:
           pip install -e .
       - name: Sphinx build
         run: |
-          sphinx-build docs docs_build
+          sphinx-build docs _build
       - name: Deploy to GitHub Pages
         uses: peaceiris/actions-gh-pages@v3
         if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
         with:
           publish_branch: gh-pages
           github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: docs_build/
+          publish_dir: _build/
           force_orphan: true
