### 2024-09-27 BattModels/asimtools#45 @mkphuthi
Great. I understand now. I have edited the ReadMe to point immediately to the [Developing Custom Asimmodules](https://eeg.engin.umich.edu/asimtools/asimplify.html) page where I have tried to make clear the advantages of using ASIMTools over the traditional approach in a problem-solution structure.

@mkphuthi
