### 2024-10-09 a551a9d
Update README.md
Minor typos
