+++ _sources/asimtools.asimmodules.lammps.rst.txt
@@ -1,0 +1,20 @@
+asimtools.asimmodules.lammps package
+====================================
+
+
+
+asimtools.asimmodules.lammps.lammps module
+------------------------------------------
+
+.. automodule:: asimtools.asimmodules.lammps.lammps
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: asimtools.asimmodules.lammps
+   :members:
+   :undoc-members:
+   :show-inheritance:
