+++ asimtools/asimmodules/phonopy/phonon_bands_and_dos.py
@@ -19,16 +19,43 @@ def phonon_bands_and_dos(
     calc_env_id: str,
     process_env_id: str,
     supercell: ArrayLike = [10,10,10],
-    distance: float = 0.02,
-    phonopy_save_path: str = 'phonopy_save.yaml',
-    paths: Optional[Sequence] = None,
-    labels: Optional[Sequence] = None,
+    distance: Optional[float] = 0.02,
+    phonopy_save_path: Optional[str] = 'phonopy_save.yaml',
+    paths: Optional[ArrayLike] = None, 
+    labels: Optional[ArrayLike] = None,
     use_seekpath: Optional[bool] = True,
-    npoints: int = 51,
-    mesh: Union[ArrayLike,float] = [20, 20, 20],
-    t_max: float = 1000,
+    npoints: Optional[int] = 51,
+    mesh: Optional[Union[ArrayLike,float]] = [20, 20, 20],
 ) -> Dict:
+    """Calculates band structure and density of states using phonopy.
 
+    :param image: Image specification. See :ref:`asimtools.utils.get_image`.
+    :type image: Dict
+    :param calc_id: calc_id of the calculator to use.
+    :type calc_id: str
+    :param calc_env_id: env_id of the calculator to use.
+    :type calc_env_id: str
+    :param process_env_id: env_id for pre- and post-processing.
+    :type process_env_id: str
+    :param supercell: supercell of specified image, defaults to [10,10,10]
+    :type supercell: ArrayLike, optional
+    :param distance: Distance in phonopy, defaults to 0.02
+    :type distance: float, optional
+    :param phonopy_save_path: Where to put the phonopy yaml, defaults to 'phonopy_save.yaml'
+    :type phonopy_save_path: str, optional
+    :param paths: Paths to display in band  structure. Has shape Np*Nk*3 where Np is the number of disconnected paths, Kn is the number of points in each path and lastly the coordinate axes in kspace, defaults to None
+    :type paths: Optional[ArrayLike], optional
+    :param labels: Labels for each point provided in paths, defaults to None
+    :type labels: Optional[ArrayLike], optional
+    :param use_seekpath: Automatically generate a suggested path, defaults to True
+    :type use_seekpath: Optional[bool], optional
+    :param npoints: Number of points for band path, defaults to 51
+    :type npoints: int, optional
+    :param mesh: Size of kpoint mesh for DOS either as integer or array of length 3, defaults to [20, 20, 20]
+    :type mesh: Union[ArrayLike,float], optional
+    :return: result
+    :rtype: Dict
+    """
     phonopy_save_path = str(Path(phonopy_save_path).resolve())
     sim_input = {
         'asimmodule': 'workflows.chained',
