### 2024-10-15 BattModels/asimtools#50 @HaoZeke
[JOSS REVIEW] Please make a release
Part of https://github.com/openjournals/joss-reviews/issues/7085

I see there is a PyPI version, but it is set at v0.0.0

Please:
- Make a new tag (v0.0.1 perhaps)
- Make a release
- Upload to PyPI

Additionally, for JOSS we will need to have a Zenodo archive of this version as well (to be set as the version of record).
