### 2024-10-16 70ada40
Added vasp example, change job array command since slurm changed syntax
