+++ examples/external/VASP/singlepoint_sim_input.yaml
@@ -1,0 +1,9 @@
+asimmodule: singlepoint
+workdir: vasp_results
+env_id: inline
+args:
+  calc_id: vasp
+  image:
+    name: Na
+    builder: bulk
+  properties: ['energy', 'forces']
