+++ examples/external/VASP/calc_input.yaml
@@ -1,0 +1,11 @@
+vasp_PBE:
+  name: Vasp
+  module: ase.calculators.vasp
+  precommands:
+    - module load myvasp
+  args:
+    prec: Normal
+    xc: PBE
+    lreal: False
+    command: srun vasp_std
+
