### 2024-10-16 306479f
update version, add pypi+release action
