### 2024-10-16 97d058d
update version, add pypi+release action
