### 2024-10-16 f28321b
add readme to setup.py
