+++ setup.py
@@ -39,4 +39,6 @@ setup(
         "mp-api",
     ],
     zip_safe=True,
+    long_description=open('README.md').read(),
+    long_description_content_type='text/markdown',
 )
