+++ testpyproject.toml
@@ -1,46 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
-name = "asimtools"
-version = "0.0.1"
-authors = [
-  { name="Keith Phuthi", email="mkphuthi@gmail.com" },
-]
-maintainers = [
-  { name="Keith Phuthi", email="mkphuthi@gmail.com" },
-]
-description = "A lightweight python package for managing and running atomic simulation workflows"
-readme = "README.md"
-requires-python = ">=3.9"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: MacOS",
-    "Operating System :: Unix"
-]
-dependecies = [
-    "pandas",
-    "pyyaml",
-    "pymatgen",
-    "ase>=3.22.1", 
-    # Recommended to use the master branch of the ASE gitlab but only 
-    # required for lammps to handle masses correctly as far as we can tell
-    "colorama",
-    "myst-parser",
-    "sphinx",
-    "pytest",
-    "phonopy",
-    "seekpath",
-    "mp-api",
-]
-
-[project.scripts]
-asim-run = "asimtools.scripts.asim_run:main"
-asim-execute = "asimtools.scripts.asim_execute:main"
-asim-check = "asimtools.scripts.asim_check:main"
-
-[project.urls]
-Homepage = "https://github.com/BattModels/asimtools"
-Issues = "https://github.com/BattModels/asimtools/issues"
