### 2024-10-16 BattModels/asimtools#50 @mkphuthi
Hi @HaoZeke

I've made the suggested changes. See:

Release: https://github.com/BattModels/asimtools/releases/tag/v0.0.1
PyPI distribution: https://pypi.org/manage/project/asimtools/release/0.0.1/
Zenodo archive: https://zenodo.org/records/13942780

Let me know if there is anything else necessary.
