### 2024-10-16 BattModels/asimtools#50 @HaoZeke
Thanks, that should be all for this issue.
