+++ _sources/README.md.txt
@@ -1,0 +1,189 @@
+<!-- <img src="../figures/logo.png" alt="drawing" width="150"/> -->
+<!-- [![code coverage](https://img.shields.io/codecov/c/gh/materialsproject/jobflow/main)](https://codecov.io/gh/materialsproject/jobflow/) -->
+<!-- [![pypi version](https://img.shields.io/pypi/v/jobflow?color=blue)](https://pypi.org/project/jobflow/) -->
+<!-- ![supported python versions](https://img.shields.io/pypi/pyversions/jobflow) -->
+<!-- [![DOI](https://joss.theoj.org/papers/10.21105/joss.05995/status.svg)](https://doi.org/10.21105/joss.05995) -->
+
+[Documentation](https://battmodels.github.io/asimtools/) | [GitHub](https://github.com/BattModels/asimtools)
+
+# Atomic SIMulation Tools
+
+This package is a lightweight workflow and simulation manager for reproducible
+atomistic simulations that can be transferred across environments, calculators
+and structures on Unix systems. By using in-built or user-defined asimmodules
+and utilities, users can run/build their own simulation recipes and
+automagically scale them locally or on slurm based clusters. The core idea is
+to separate the dependence of the atomistic potential/calculator and the
+simulations steps thereby allowing the same simulation to be run with multiple
+calculators/codes and the same calculator to be used for multiple simulation
+parameters without altering simulation code. Input and output files follow a
+simple consistent file structure and format so that consistent analysis
+pipelines can be used across users. For a concrete example of how ASIMTools
+achieves this, see the [Developing Custom Asimmodules](https://eeg.engin.umich.edu/asimtools/asimplify.html) page.
+
+## Developer philosophy
+The goal of asimtools is to push all the complexity of workflow management,
+best-practices, file management etc. into the backend such that the everyday
+user only has to handle input files for existing workflows and asimmodule files
+for workflows they want to implement.This allows the user to focus on doing the
+science and designing experiments. The following are the guiding principles for
+how asimtools should work:
+
+- Asimmodules should resemble boilerplate ASE code as much as possible.
+- Asimmodules should avoid explicitly depending on a calculator
+- Asimmodules should not explicitly depend on the context/environment in which 
+  they run
+- It should be easy to debug individual asimmodules/parts in large workflows.
+  In addition, it should always be possible to debug/resubmit jobs without
+  using asimtools.
+- Input file structure and format should be standard across all asimmodules. In
+  addition all input parameters should match how they would like without
+  asimtools i.e. do not provide an API!
+- Job progress tracking must be incorporated for easy debugging
+- Best practices should be built-in e.g. if multiple jobs of the same slurm
+  context are submitted simulataneously, it must be a job array.
+
+## Philosophy on User Experience
+The philosophy is to build "simulations" using building blocks of asimmodules.
+Asimmodules are nothing but Python functions that return a dictionary, anything
+can be done in the function code. These asimmodules can be as
+complicated/efficient as you make them using any external packages you want and
+can be optimized with time but can still be run within the framework. This
+allows a test friendly way to transition from say a tutorial on the
+ASE/pymatgen website to an asimtools asimmodule. So while complicated wrappers
+are discouraged, they would still work as long as the asimmodule works. The
+benefit out of the box is that you can make your asimmodule independent of
+calculator or input structures and submit them easily.
+
+We also aim to provide a standard set of robust and efficient simulation
+protocols as we develop. You can see all the implemented workflows provided
+with the package in the examples directory and modify them to your application.
[…]
