### 2024-10-18 #7085 @HaoZeke
@mkphuthi could you make a new release (zenodo only should be fine) with the updated bibliography?
