### 2024-10-18 2df95ed
Update pyproject.toml
Update bibliography
