+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "asimtools"
 description = "A lightweight python package for managing and running atomic simulation workflows"
-version = "0.0.1"
+version = "0.0.2"
 readme = "README.md"
 license = { text = "MIT" }
 authors = [
