### 2024-10-18 2a8da8d
Update _version.py
Update bibliography
