+++ asimtools/_version.py
@@ -1,4 +1,4 @@
 # See Python packaging guide
 # https://packaging.python.org/guides/single-sourcing-package-version/
 
-__version__ = "0.0.1"
+__version__ = "0.0.2"
