### 2024-10-18 #7085 @HaoZeke
@mkphuthi sorry but the zenodo release is v2 but it should be v0.0.2, please update / reissue it.
