### 2024-10-19 #7085 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
