### 2024-10-20 #7085 @Kevin-Mattheus-Moerman
@mkphuthi thanks for editing this one. Minor note, when setting the DOI, focus on just the DOI, so trim off the rest of the URL.
