### 2024-10-20 #7085 @Kevin-Mattheus-Moerman
@mkphuthi as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. These are some final checks I am doing:

- [ ] Software license tag listed here matches a tagged release

- [ ]
